+++ app/safe/internal/server/route/base/validation/validation.go
@@ -14,9 +14,12 @@ import (
 	"io"
 	"net/http"
 
+	ioState "github.com/vmware-tanzu/secrets-manager/app/safe/internal/state/io"
 	"github.com/vmware-tanzu/secrets-manager/core/audit/journal"
 	"github.com/vmware-tanzu/secrets-manager/core/constants/audit"
+	"github.com/vmware-tanzu/secrets-manager/core/constants/val"
 	"github.com/vmware-tanzu/secrets-manager/core/entity/v1/data"
+	"github.com/vmware-tanzu/secrets-manager/core/env"
 	log "github.com/vmware-tanzu/secrets-manager/core/log/std"
 	"github.com/vmware-tanzu/secrets-manager/core/validation"
 )
@@ -65,4 +68,37 @@ func IsSentinel(
 	}
 
 	return false, responder
+}
+
+// CheckDatabaseReadiness checks if the database is ready for use.
+//
+// This function verifies the readiness of the database, specifically for
+// PostgreSQL mode. If the database is not initialized when PostgreSQL
+// mode is enabled, it returns an error response.
+//
+// Parameters:
+//   - cid: A string representing the context or correlation ID for logging.
+//   - w: An http.ResponseWriter to write the HTTP response.
+//
+// Returns:
+//   - bool: true if the database is ready, false otherwise.
+//
+// Side effects:
+//   - Writes an HTTP 503 (Service Unavailable) status and response body
+//     if the database is not ready.
+//   - Logs information about the database status.
+func CheckDatabaseReadiness(cid string, w http.ResponseWriter) bool {
+
+	// If postgres mode enabled and db is not initialized, return error.
+	if env.BackingStoreForSafe() == data.Postgres && !ioState.PostgresReady() {
+		w.WriteHeader(http.StatusServiceUnavailable)
+		_, err := io.WriteString(w, val.NotOk)
+		if err != nil {
+			log.ErrorLn(&cid, "error writing response", err.Error())
+		}
+		log.InfoLn(&cid, "Secret: Database not initialized")
+		return false
+	}
+
+	return true
 }