+++ app/safe/internal/server/route/secret/secret.go
@@ -11,6 +11,8 @@
 package secret
 
 import (
+	ioState "github.com/vmware-tanzu/secrets-manager/app/safe/internal/state/io"
+	"github.com/vmware-tanzu/secrets-manager/core/env"
 	"io"
 	"net/http"
 	"time"
@@ -117,6 +119,24 @@ func Secret(cid string, r *http.Request, w http.ResponseWriter) {
 	appendValue := sr.AppendValue
 	notBefore := sr.NotBefore
 	expiresAfter := sr.Expires
+
+	// The next check is only for non-vsecm-safe workloads.
+	if len(workloadIds) == 0 || workloadIds[0] != "vsecm-safe" {
+		log.TraceLn(&cid, "Secret: workloadIds is empty or not vsecm-safe")
+
+		// If postgres mode enabled and db is not initialized, return error.
+		if env.BackingStoreForSafe() == entity.Postgres && !ioState.PostgresReady() {
+			w.WriteHeader(http.StatusServiceUnavailable)
+			_, err := io.WriteString(w, val.NotOk)
+			if err != nil {
+				log.ErrorLn(&cid, "error writing response", err.Error())
+			}
+			log.InfoLn(&cid, "Secret: Database not initialized")
+			return
+		}
+	} else {
+		log.TraceLn(&cid, "Secret: vsecm-safe workload detected")
+	}
 
 	if len(workloadIds) == 0 && encrypt {
 		httq.SendEncryptedValue(cid, value, j, w)