+++ app/safe/internal/state/secret/collection/populate.go
@@ -73,8 +73,11 @@ func PopulateSecrets(cid string) error {
 			log.ErrorLn(&cid, "populateSecrets:error", err.Error())
 		}
 	case data.Postgres:
-		// TODO: implement me.
-		log.WarnLn(&cid, "populateSecrets: postgres initial secrets population is not implemented yet.")
+		err := populateSecretsFromPostgresqlDb(cid)
+		populated = err == nil
+		if err != nil {
+			log.ErrorLn(&cid, "populateSecrets:error", err.Error())
+		}
 	case data.Kubernetes:
 		panic("implement kubernetes store")
 	case data.AwsSecretStore: